refactor : SignupEmail 리팩토링 및 styles 분리
